+++ lib/src/markdown_content.dart
@@ -8,10 +8,7 @@ import 'dart:io';
 import 'package:markdown/markdown.dart';
 import 'package:meta/meta.dart';
 import 'package:html/parser.dart' as html_parser;
-import 'package:path/path.dart' as p;
 import 'package:source_span/source_span.dart';
-
-import 'model.dart';
 
 /// The extracted content of a markdown file.
 class ExctractedMarkdownContent {
@@ -48,48 +45,6 @@ List<T> _unique<T>(Iterable<T> l) => l.toSet().toList();
 Future<ExctractedMarkdownContent> scanMarkdownFileContent(File file) async {
   final text = await file.readAsString();
   return scanMarkdownText(text, file.uri);
-}
-
-/// Analyze a markdown file and return a composite suggestion.
-Future<Suggestion> analyzeMarkdownFile(File file, {String pkgDir}) async {
-  final fileName = p.basename(file.path);
-  final relativePath = p.relative(file.path, from: pkgDir);
-  final analysis = await scanMarkdownFileContent(file);
-
-  final checked = await checkLinks(analysis.images);
-  // TODO: warn about relative image URLs
-  // TODO: warn about insecure links
-  // TODO: warn about relative links
-  // TODO: consider checking whether the URL exists and returns HTTP 200.
-
-  final issues = <String>[];
-  var score = 0.0;
-  if (checked.unparsed.isNotEmpty) {
-    final count = checked.unparsed.length;
-    final first = checked.unparsed.first.url;
-    score += count;
-    final pluralize = count == 1 ? 'link' : 'links';
-    issues.add('Unable to parse $count image $pluralize (e.g. `$first`).');
-  }
-  if (checked.insecure.isNotEmpty) {
-    final count = checked.insecure.length;
-    final first = checked.insecure.first.url;
-    score += count * 2;
-    final pluralize = count == 1 ? 'link is' : 'links are';
-    issues.add(
-        '$count image $pluralize insecure (e.g. `$first`), use `https` URLs instead.');
-  }
-  if (score > 0.0) {
-    return Suggestion.hint(
-      SuggestionCode.markdownContent,
-      'Update `$fileName`.',
-      issues.join('\n\n'),
-      score: score,
-      file: relativePath,
-    );
-  }
-
-  return null;
 }
 
 Future<Links> checkLinks(List<Link> links) async {