+++ test/markdown_content_test.dart
@@ -4,7 +4,6 @@
 
 import 'dart:io';
 
-import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 
 import 'package:pana/src/markdown_content.dart';
@@ -17,8 +16,6 @@ void main() {
       'images': ['https://example.com/path.jpg', 'img/path.jpg'],
       'links': ['mailto:hello@example.com'],
     });
-    final suggestion = await analyzeMarkdownFile(file);
-    expect(suggestion, isNull);
   });
 
   test('bad_images.md', () async {
@@ -31,16 +28,6 @@ void main() {
         'gopher://example.com/logo.png',
       ],
       'links': [],
-    });
-    final suggestion = await analyzeMarkdownFile(file);
-    expect(suggestion.toJson(), {
-      'code': 'markdown.content',
-      'level': 'hint',
-      'title': 'Update `bad_images.md`.',
-      'description':
-          '2 image links are insecure (e.g. `http://example.com/logo.png`), use `https` URLs instead.',
-      'file': p.normalize('test/markdown_content/bad_images.md'),
-      'score': 4.0,
     });
   });
 }